+++ qe/user/__init__.py
@@ -11,6 +11,7 @@ class User(API):
     from qe.user.exchange import list_exchange_apis
     from qe.user.trading import get_master_orders
     from qe.user.trading import get_master_order_detail
+    from qe.user.trading import get_master_order_detail_by_client_order_id
     from qe.user.trading import get_order_fills
     from qe.user.trading import get_tca_analysis
     from qe.user.trading import create_master_order